Created files main.c, ft_putchar.c and rush01.c in the folder rush01.
Added the function main, to run the program, version of the rush 01.

Added the function ft_putchar is to print the values on the screen, version of the rush 01.

Added the functions rush and verify, function rush is for to create the square and the verify is to insert the correct value on the screen, version of the rush 01.